https://github.com/visionmedia/express/wiki/Migrating-from-3.x-to-4.x
Express 4 no longer has connect as a dependency. This means that ALL of the bundled middleware (except static) is no longer available on the express module. Each of these middlewares is available as modules referenced below. This change was made to allow for these middleware to receive fixes, updates, and releases without impacting express release cycles and vice versa.